relocate some structs to a new progressive_status_types.go file
